Fixed a silly double to int bug that was screwing up Enu ranges set in fSampleSettings.SetEnuRange()

+++ src/Utils/GeneralUtils.cxx
@@ -32,7 +32,7 @@ std::string GeneralUtils::IntToStr(int val) {
   return ss.str();
 };
 
-std::string GeneralUtils::DblToStr(int val) {
+std::string GeneralUtils::DblToStr(double val) {
   std::ostringstream ss;
   ss << val;
   return ss.str();
